Move cardinality constraint test to dedicated module
