+++ tests/constraints/test_constraints_continuous.py
@@ -1,24 +1,11 @@
 """Test for imposing continuous constraints."""
 
-import warnings
-
 import numpy as np
-import pandas as pd
 import pytest
 from pytest import param
 
 from baybe.constraints import ContinuousLinearConstraint
-from baybe.constraints.continuous import ContinuousCardinalityConstraint
-from baybe.parameters.numerical import NumericalContinuousParameter
-from baybe.recommenders.pure.bayesian.base import BayesianRecommender
-from baybe.recommenders.pure.bayesian.botorch import BotorchRecommender
-from baybe.recommenders.pure.nonpredictive.sampling import RandomRecommender
-from baybe.searchspace import SearchSpace
-from baybe.targets.numerical import NumericalTarget
 from tests.conftest import run_iterations
-from tests.constraints.test_cardinality_constraint_continuous import (
-    _validate_cardinality_constrained_batch,
-)
 
 
 @pytest.mark.parametrize("parameter_names", [["Conti_finite1", "Conti_finite2"]])
@@ -79,40 +66,6 @@ def test_inequality3(campaign, n_iterations, batch_size):
     print(res)
 
     assert (1.0 * res["Conti_finite1"] + 3.0 * res["Conti_finite2"]).le(0.301).all()
-
-
-@pytest.mark.parametrize("recommender", [RandomRecommender(), BotorchRecommender()])
-def test_cardinality_constraint(recommender):
-    """Cardinality constraints are taken into account by the recommender."""
-    MIN_CARDINALITY = 4
-    MAX_CARDINALITY = 7
-    BATCH_SIZE = 10
-
-    parameters = [NumericalContinuousParameter(str(i), (0, 1)) for i in range(10)]
-    constraints = [
-        ContinuousCardinalityConstraint(
-            [p.name for p in parameters], MIN_CARDINALITY, MAX_CARDINALITY
-        )
-    ]
-    searchspace = SearchSpace.from_product(parameters, constraints)
-
-    if isinstance(recommender, BayesianRecommender):
-        objective = NumericalTarget("t", "MAX").to_objective()
-        measurements = pd.DataFrame(searchspace.continuous.sample_uniform(2))
-        measurements["t"] = np.random.random(len(measurements))
-    else:
-        objective = None
-        measurements = None
-
-    with warnings.catch_warnings(record=True) as w:
-        recommendation = recommender.recommend(
-            BATCH_SIZE, searchspace, objective, measurements
-        )
-
-    # Assert that the constraint conditions hold
-    _validate_cardinality_constrained_batch(
-        recommendation, searchspace.continuous, BATCH_SIZE, w
-    )
 
 
 @pytest.mark.slow
